requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
transformers==4.38.2
torch==2.2.1
elasticsearch==8.12.1
scikit-learn==1.4.2
python-Levenshtein==0.23.0
gunicorn==21.2.0
Werkzeug==3.0.1 